fix: reorganize file locations
